backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
langchain-community==0.0.13
langchain-google-genai==0.0.5
chromadb==0.4.18
pypdf==3.17.1
python-dotenv==1.0.0
google-generativeai==0.3.1 